backend/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
pymongo>=4.5.0
python-dotenv>=1.0.0
passlib>=1.7.4
pyjwt>=2.8.0
bcrypt>=4.0.1
httpx>=0.25.0
python-multipart>=0.0.6
pandas>=2.0.0
yfinance>=0.2.27
requests>=2.31.0
matplotlib>=3.10.3
scipy>=1.15.0,<1.16.0
seaborn>=0.13.2
pytest>=7.4.0

# You'll need to install ibapi from Interactive Brokers' website
# https://interactivebrokers.github.io/ 
# To install all requirements, run:
# pip install -r requirements.txt
